server/requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click @ file:///tmp/build/80754af9/click_1621604852318/work
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
dlib==19.22.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
jwcrypto==0.9
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
numpy==1.20.3
oauth2client==4.1.3
opencv-python==4.5.2.52
Pillow==8.2.0
pipreqs==0.4.10
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
pyrebase5==5.0.1
python-jwt==3.3.0
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.5
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
